Add laying down on activating bed menu
While the bed menu is open, the player will be placed lying down on the
bed without actually sleeping, as the default sleeping mechanism is
conditional and doesn't allow a menu to be opened at the same time. This
implementation was done using the GSit API in order to have the player
fake laying down. The player remains laying down after a teleport, so
they have to get up manually.