Fix direction from vector flipped
Bed facing direction refers to the opposite side, so the vector
conversion to direction was incorrect. Vector conversion has been
flipped alongside the references to bed facing direction.

This issue isn't an issue for this one case, but if other things need to
use this conversion, then it'll be flipped.