Add handling for teleport cost
If the cost is set and the player doesn't have enough of the item,
refuse to teleport. Otherwise if the player has the required cost, take
it from their inventory as soon as the teleport happens.

This change also adds a new callback for if the world is not found.